Refactor tenant picker to a seperate view controller (SP-674)
